Powoli do przodu - obsługa Calc.xxxx

+++ src/main/java/Calc/Calculator.java
@@ -2,23 +2,23 @@ package Calc;
 
 public class Calculator {
 
-    public double add(int a, int b) {
+    public double add(double a, double b) {
         return a+b;
     }
 
-    public double subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a-b;
     }
 
-    public double multiply (int a, int b) {
+    public double multiply (double a, double b) {
         return a*b;
     }
 
-    public double divide (int a, int b) {
+    public double divide (double a, double b) {
         return a/b;
     }
 
-    public double MathPow(int a, int b) {
+    public double MathPow(double a, double b) {
         return Math.pow(a,b);
     }
 }
